bug fix vercel deployment / undefined properties
